Figure to test with.
I had him rigged and posed earlier, but that was prior to finishing the
hands, so I removed the rig and finished it. I found a tutorial that
makes it rather easy to rig it, so, I had it rigged in about 30 seconds.

It looks a bit off in some places, but it's a good mesh to start with
and base other models off. Clothes and armor and stuff will obscure the
shape of the base mesh a bit.

At its lowest subdivision, it's 654 polygons. At one level of
subdivision, it's around 2500. And that's without facial features and
fine details.